+++ packages/python-packages/api-stub-generator/apistub/nodes/_class_node.py
@@ -161,9 +161,9 @@ class ClassNode(NodeEntityBase):
         docstring = getattr(self.obj, "__doc__")
         if docstring:
             docstring_parser = DocstringParser(docstring)
-            for var in docstring_parser.find_args("ivar"):
+            for key, var in docstring_parser.ivars.items():
                 ivar_node = VariableNode(
-                    self.namespace, self, var.argname, var.argtype, None, True
+                    self.namespace, self, key, var.argtype, None, True
                 )
                 self.child_nodes.append(ivar_node)