+++ packages/python-packages/api-stub-generator/CHANGELOG.md
@@ -5,6 +5,8 @@ Updated type parsing to properly handle cases in which type
   info wraps onto a second line.
 Fixed issue where ivar type was not properly parsed from
   the docstring.
+Fixed issue where typehint parsing was too strict and required
+  spaces.
 
 ## Version 0.2.5 (Unreleased)
 Fixed bug in which kwargs were duplicated in certain instances.